requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
beautifulsoup4==4.9.3
Flask==2.0.1
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
gunicorn==20.1.0
huggingface-hub==0.0.12
PyJWT==2.0.0
pymongo[srv]==3.12.0
pyparsing==2.4.7
python-dotenv==0.19.0
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
tokenizers==0.10.3
toml==0.10.2
torch==1.9.1+cu111
torchaudio==0.9.1
torchvision==0.10.1+cu111
tqdm==4.61.1
transformers==4.8.2
